# api-management/api-documentation.yaml
openapi: '3.0.0'
info:
  version: 1.0.0
  title: Documentation API du projet money
  description: Décrit la structure de l'API du projet money de l'équipe NALT
servers:
    - url: http://localhost:3000/api
      description: Serveur de développement de l'API
paths:
...
  /categories:
    get:
      summary: Retourne une liste de catégories.
      parameters:
        - name: idCategorie
          description: Identifiant de la catégorie.
          in: query
          required: false
          schema:
            type: integer

      security:
        - bearerAuth: [ ]
      tags:
        - Categories
      description: Retourne une liste de catégories avec les champs idCategorie, nomCategorie, dateHeureCreation et dateHeureMAJ.
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '500':
          $ref: '#/components/responses/InternalServerError'


    post:
      summary: Catégorie.
      security:
        - bearerAuth: [ ]
      tags:
        - Categories
      description: Crée une nouvelle catégorie avec le champ nomCategorie.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nomCategorie:
                  type: string
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '201':
          $ref: '#/components/responses/Created'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '500':
          $ref: '#/components/responses/InternalServerError'     
  /categories/{idCategorie}:
    get:
      summary: Retourne une catégorie par son identifiant.
      security:
        - bearerAuth: [ ]
      tags:
        - Categories
      description: Retourne une catégorie avec les champs idCategorie, nomCategorie, dateHeureCreation et dateHeureMAJ.
      parameters:
        - name: idCategorie
          description: Identifiant de la catégorie.
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '500':
          $ref: '#/components/responses/InternalServerError'


    put:
      summary: Met à jour une catégorie.
      security:
        - bearerAuth: [ ]
      tags:
        - Categories
      description: Met à jour une catégorie avec les champs nomCategorie, dateHeureCreation et dateHeureMAJ.
      parameters:
        - name: idCategorie
          description: Identifiant de la catégorie.
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nomCategorie:
                  type: string
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Supprime une catégorie.
      security:
        - bearerAuth: [ ]
      tags:
        - Categories
      description: Supprime une catégorie à partir de son identifiant.
      parameters:
        - name: idCategorie
          description: Identifiant de la catégorie.
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '204':
          '$ref': '#/components/responses/NoContent'
        '303':
          $ref: '#/components/responses/SeeOther'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /categories/{idCategorie}/sous-categories:
    get:
      summary: Retourne une liste de sous-catégories.
      parameters:
        - name: idCategorie
          description: Identifiant de la catégorie.
          in: path
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: [ ]
      tags:
        - Sous-categories
      description: Retourne une liste de sous-catégories avec les champs idSousCategorie, nomSousCategorie, idCategorie, dateHeureCreation et dateHeureMAJ.
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '500':
          $ref: '#/components/responses/InternalServerError'


    post:
      summary: Crée une nouvelle sous-catégorie.
      parameters:
        - name: idCategorie
          description: Identifiant de la catégorie
          in: path
          required: true
          schema:
            type: integer

      security:
        - bearerAuth: [ ]
      tags:
        - Sous-categories
      description: Crée une nouvelle sous-catégorie avec les champs nomSousCategorie et idCategorie.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nomSousCategorie:
                  type: string
                idCategorie:
                  type: integer
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '201':
          $ref: '#/components/responses/Created'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /categories/{idCategorie}/sous-categories/{idSousCategorie}:
    put:
      summary: Met à jour une sous-catégorie.
      security:
        - bearerAuth: [ ]
      tags:
        - Sous-categories
      description: Met à jour une sous-catégorie avec le champ nomSousCategorie.
      parameters:
        - name: idSousCategorie
          description: Identifiant de la sous-catégorie.
          in: path
          required: true
          schema:
            type: integer
        - name: idCategorie
          description: Identifiant de la catégorie.
          in: path
          required: true
          schema:
            type: integer

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nomSousCategorie:
                  type: string
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '500':
          $ref: '#/components/responses/InternalServerError'


    delete:
      summary: Supprime une sous-catégorie.
      security:
        - bearerAuth: [ ]
      tags:
        - Sous-categories
      description: Supprime une sous-catégorie à partir de son identifiant.
      parameters:
        - name: idSousCategorie
          description: Identifiant de la sous-catégorie.
          in: path
          required: true
          schema:
            type: integer
        - name: idCategorie
          description: Identifiant de la catégorie.
          in: path
          required: true
          schema:
            type: integer

      responses:
        '200':
          $ref: '#/components/responses/Success'
        '204':
          '$ref': '#/components/responses/NoContent'
        '303':
          $ref: '#/components/responses/SeeOther'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  ###################################################################################

  # components
components:
...
  schemas:    
...
    Error:
      description: Error response
      type: object
      properties:
        code:
          type: string
        message:
          type: string
    
    # schema for the success responses
    Success:
      type: object
      properties:
        message:
          type: string
        data:
          type: object
          additionalProperties: true
      
  # responses
  responses:
...
    Success:
      description: Success
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Success'
    # 201 Created
    Created:
      description: Created
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Success'
    # 204 No Content
    NoContent:
      description: No Content
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Success'
    # 303 See Other
    SeeOther:
      description: See Other
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Success'     
    # 400 Bad Request
    BadRequestError:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    # 401 Unauthorized
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    # 403 Forbidden
    ForbiddenError:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    # 404 Not Found
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'  
    # 406 Not Acceptable
    NotAcceptableError:
      description: Not Acceptable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    # 409 Conflict
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  # securitySchemes
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

# security
security:
  - bearerAuth: []